refactor: extract Trends header/empty/formatting from TrendsView (#80, pt 1)
First of two steps decomposing TrendsView.swift (the 1648-line file_length
offender, type_body_length 1305). This part extracts the low-coupling,
presentational pieces along clean seams; the chart/hero/fullscreen cluster
(which carries interaction state) follows in pt 2.

Extracted into sibling files (matching the Trends folder's existing
component-per-file pattern):
  - LandscapeHostingController.swift — the fullscreen-rotation representable
    (pure infra, zero coupling)
  - TrendsFormatting.swift — value / goalLabel / periodSubtitle as pure
    static funcs, now unit-tested (TrendsFormattingTests, 7 cases)
  - TrendsMetricChip.swift — the metric selector chip (presentational; reports
    taps via onTap)
  - TrendsRangeTabs.swift — M/6M/Y/All selector (presentational; reports via
    onSelect)
  - TrendsEmptyState.swift — cold-start empty state (reports CTA via
    onSelectTab)

The extracted controls are pure/presentational: TrendsView keeps ownership of
state and passes values + closures down, so behavior is unchanged. TrendsView
shrinks 1691 → 1502 lines. Thresholds unchanged — file_length is still pinned
by the remaining chart/hero/fullscreen code until pt 2.

Verified: make lint (0 violations, --strict), make build, make test
(Baseline-CI: 266 logic incl. new TrendsFormattingTests + 13 UI, all green).
Snapshot tests fail identically on clean main (pre-existing simulator/Xcode
env drift, excluded from CI — not a regression from this change).
